Check permissions for Like module

+++ controllers/like/LikeController.php
@@ -7,6 +7,7 @@
 
 namespace humhub\modules\rest\controllers\like;
 
+use humhub\modules\content\models\Content;
 use humhub\modules\rest\components\BaseController;
 use humhub\modules\rest\definitions\LikeDefinitions;
 use humhub\modules\like\models\Like;
@@ -18,9 +19,21 @@ class LikeController extends BaseController
 
     public function actionFindByRecord()
     {
+        $contentFilter = [
+            'object_model' => Yii::$app->request->get('model'),
+            'object_id' => (int)Yii::$app->request->get('pk'),
+        ];
+        $content = Content::findOne($contentFilter);
+        if ($content === null) {
+            return $this->returnError(404, 'Content not found!');
+        }
+        if (!$content->canView()) {
+            return $this->returnError(403, 'You cannot view this content!');
+        }
+
         $results = [];
         $query = Like::find();
-        $query->andWhere(['object_model' => Yii::$app->request->get('model'), 'object_id' => (int)Yii::$app->request->get('pk')]);
+        $query->andWhere($contentFilter);
         $query->orderBy(['created_at' => SORT_DESC]);
 
         $pagination = $this->handlePagination($query);
@@ -39,6 +52,9 @@ class LikeController extends BaseController
         if ($like === null) {
             return $this->returnError(404, 'Like not found!');
         }
+        if (!$like->canRead()) {
+            return $this->returnError(403, 'You cannot read this content!');
+        }
 
         return LikeDefinitions::getLike($like);
     }
@@ -48,6 +64,9 @@ class LikeController extends BaseController
         $like = Like::findOne(['id' => $id]);
         if ($like === null) {
             return $this->returnError(404, 'Like not found!');
+        }
+        if (!$like->canDelete()) {
+            return $this->returnError(403, 'You cannot delete this content!');
         }
 
         if ($like->delete()) {
